@@ -1,8 +1,7 @@
 source 'https://rubygems.org'
 
-gem 'rails',          '4.2.7'
+gem 'rails',          '5.0.1'
 gem 'responders'
-gem 'oj'
 gem 'pg'
 gem 'pg_search'
 gem 'bcrypt'
@@ -13,12 +12,14 @@ gem 'gcoder'
 gem 'puma'
 gem 'jquery-rails'
 gem 'bootstrap-sass', '~> 3.3.3'
-gem 'sass-rails',     '~> 5.0.1'
-gem 'therubyracer',   platforms: :ruby
-gem 'uglifier'
-#gem 'skylight',       '~> 0.6.0'
-#gem 'honeybadger',    '~> 2.0.6'
+gem 'sass-rails',     '~> 5.0'
+gem 'uglifier',       '>= 1.3.0'
 gem 'stripe',         git: 'https://github.com/stripe/stripe-ruby'
+
+gem 'capistrano',         '~> 3.3.5'
+gem 'capistrano-rails'
+gem 'capistrano-bundler'
+gem 'capistrano-rvm'
 
 # Crawler Junk
 gem 'excon',         require: false
@@ -26,20 +27,22 @@ gem 'amatch',        require: false
 gem 'stringex',      require: false
 gem 'nokogiri',      require: false
 gem 'unicode_utils', require: false
-gem 'aws-sdk',       '~> 2'
+gem 'aws-sdk',       '~> 2', require: false
 
 group :development do
-  gem 'capistrano',         '~> 3.3.5'
-  gem 'capistrano-rails'
-  gem 'capistrano-bundler'
-  gem 'capistrano-rvm'
-  gem 'quiet_assets'
   gem 'pry-rails'
-  gem 'spring'
-  gem 'spring-commands-rspec'
+  gem 'web-console',        '>= 3.3.0'
+  gem 'listen',             '~> 3.0.5'
+
+  # Spring speeds up development by keeping your application running in the
+  # background. Read more: https://github.com/rails/spring
+  # gem 'spring'
+  # gem 'spring-watcher-listen', '~> 2.0.0'
+  # gem 'spring-commands-rspec'
 end
 
 group :test, :development do
+  gem 'byebug',        platform: :mri
   gem 'rspec-rails'
   gem 'fabrication',   require: false
   gem 'awesome_print', require: 'ap'
