@@ -3,8 +3,8 @@ class API::V1::DatasetsController < API::V1::APIController
     @query = query(:datasets)
 
     respond_to do |format|
-      format.csv { render text: @query.as_csv }
-      format.tsv { render text: @query.as_tsv }
+      format.csv { render plain: @query.as_csv }
+      format.tsv { render plain: @query.as_tsv }
       format.any(:js, :json) { render_json @query.as_json }
     end
   end
@@ -13,8 +13,8 @@ class API::V1::DatasetsController < API::V1::APIController
     @query = query(:dataset)
 
     respond_to do |format|
-      format.csv { render text: @query.as_csv }
-      format.tsv { render text: @query.as_tsv }
+      format.csv { render plain: @query.as_csv }
+      format.tsv { render plain: @query.as_tsv }
       format.zip { redirect_to @query.as_json[:result][:csv_dump] }
       format.any(:js, :json) { render_json @query.as_json }
     end
